requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
click==7.1.2
discord.py==1.3.4
dnspython==1.15.0
dnspython3==1.15.0
idna==2.10
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
mcstatus==4.0.0
multidict==4.7.6
mysql-connector-python==8.0.21
pdoc3==0.9.1
protobuf==3.13.0
pywin32-ctypes==0.2.0
six==1.15.0
websockets==8.1
yarl==1.5.1
